requirements.txt
psycopg2-binary
sentence-transformers
transformers
torch
# langchain
